restapi/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pytest==8.3.3
httpx==0.27.2
aiofiles==24.1.0
python-dotenv==1.0.1